PP/second_lab/pyproject.toml
[tool.poetry]
name = "second-lab"
version = "0.1.0"
description = ""
authors = ["Ivaschenko Nikita <ivashchenkonikal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
opencv-python = "^4.8.1.78"
numpy = "^1.26.0"
pillow = "^10.0.1"
scipy = "^1.11.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
